+++ sdk/resourcemanager/automanage/armautomanage/zz_generated_bestpractices_client.go
@@ -53,7 +53,7 @@ func NewBestPracticesClient(credential azcore.TokenCredential, options *arm.Clie
 
 // Get - Get information about a Automanage best practice
 // If the operation fails it returns an *azcore.ResponseError type.
-// Generated from API version 2021-04-30-preview
+// Generated from API version 2022-05-04
 // bestPracticeName - The Automanage best practice name.
 // options - BestPracticesClientGetOptions contains the optional parameters for the BestPracticesClient.Get method.
 func (client *BestPracticesClient) Get(ctx context.Context, bestPracticeName string, options *BestPracticesClientGetOptions) (BestPracticesClientGetResponse, error) {
@@ -83,7 +83,7 @@ func (client *BestPracticesClient) getCreateRequest(ctx context.Context, bestPra
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2021-04-30-preview")
+	reqQP.Set("api-version", "2022-05-04")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, nil
@@ -100,7 +100,7 @@ func (client *BestPracticesClient) getHandleResponse(resp *http.Response) (BestP
 
 // NewListByTenantPager - Retrieve a list of Automanage best practices
 // If the operation fails it returns an *azcore.ResponseError type.
-// Generated from API version 2021-04-30-preview
+// Generated from API version 2022-05-04
 // options - BestPracticesClientListByTenantOptions contains the optional parameters for the BestPracticesClient.ListByTenant
 // method.
 func (client *BestPracticesClient) NewListByTenantPager(options *BestPracticesClientListByTenantOptions) *runtime.Pager[BestPracticesClientListByTenantResponse] {
@@ -133,7 +133,7 @@ func (client *BestPracticesClient) listByTenantCreateRequest(ctx context.Context
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2021-04-30-preview")
+	reqQP.Set("api-version", "2022-05-04")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, nil